+++ operations/helm/charts/mimir-distributed/CHANGELOG.md
@@ -34,8 +34,9 @@ Entries should include a reference to the Pull Request that introduced the chang
 * [ENHANCEMENT] Individual mimir components can override their container images via the *.image values. The component's image definitions always override the values set in global `image` or `enterprise.image`. #10340
 * [BUGFIX] Fix calculation of `mimir.siToBytes` and use floating point arithmetics. #10044
 
-## 5.6.0-rc.0
+## 5.6.0
 
+* [ENHANCEMENT] Upgrade Mimir and GEM to [2.15.0-rc.0](https://github.com/grafana/mimir/blob/main/CHANGELOG.md#2150). #10369
 * [CHANGE] Update rollout-operator version to 0.20.0. #9995
 * [CHANGE] Remove the `track_sizes` feature for Memcached pods since it is unused. #10032
 * [FEATURE] Add support for GEM's federation-frontend. See the `federation_frontend` section in the values file. #9673